Do not escape whole string
Rather make sure to send it only properly escaped strings;
http://www.postgresql.org/docs/8.3/static/textsearch-controls.html#TEXTSEARCH-PARSING-QUERIES

to_tsquery can also accept single-quoted phrases. This is primarily
useful when the configuration includes a thesaurus dictionary that may
trigger on such phrases. In the example below, a thesaurus contains the
rule supernovae stars : sn:

[snip]

Without quotes, to_tsquery will generate a syntax error for tokens that
are not separated by an AND or OR operator.